backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.11
pgvector>=0.4.1

# Flask Services
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
requests==2.31.0

# NLP/ML
PyMuPDF>=1.26.6
python-docx==1.1.0
spacy>=3.8.11
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy>=1.24.3,<2.0
pandas>=2.1.3

# Responsible AI
aif360==0.5.0
fairlearn==0.9.0
shap==0.43.0

# Object Storage
minio==7.2.0
boto3==1.34.0

# MLOps
mlflow==2.9.2
dvc==3.40.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
python-json-logger==2.0.7
django-filter==23.5

    